Refactor course catalog, remove CoursesController BT#16815
